requirements.txt
streamlit==1.3.0
pycaret==2.3.5
pandas== 1.3.5
numpy== 1.19.5
matplotlib==3.5.1
Pillow== 8.4.0
